# specs/001-mvp-core-platform/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: QualyIT API
  description: API for QualyIT Quality Management Platform (MVP)
  version: 1.0.0
  contact:
    name: QualyIT Support
servers:
  - url: https://{tenant}.qualyit.app/api
    variables:
      tenant:
        default: demo
        description: Tenant subdomain
security:
  - bearerAuth: []
...
paths:
...
  /areas:
    get:
      tags: [Areas]
      summary: List all areas
      description: Returns hierarchical list of areas for the tenant
      parameters:
        - name: flat
          in: query
          schema:
            type: boolean
            default: false
          description: Return flat list instead of tree
      responses:
        '200':
          description: List of areas
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Area'

    post:
      tags: [Areas]
      summary: Create area
      description: Create a new organizational area
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAreaRequest'
      responses:
        '201':
          description: Area created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Area'
        '400':
          $ref: '#/components/responses/BadRequest'
  /areas/{areaId}:
    get:
      tags: [Areas]
      summary: Get area details
      parameters:
        - $ref: '#/components/parameters/AreaId'
      responses:
        '200':
          description: Area details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AreaWithStats'
        '404':
          $ref: '#/components/responses/NotFound'

    patch:
      tags: [Areas]
      summary: Update area
      parameters:
        - $ref: '#/components/parameters/AreaId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateAreaRequest'
      responses:
        '200':
          description: Area updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Area'

    delete:
      tags: [Areas]
      summary: Delete area
      description: Soft delete an area (sets is_active=false)
      parameters:
        - $ref: '#/components/parameters/AreaId'
      responses:
        '204':
          description: Area deleted
  /areas/{areaId}/stats:
    get:
      tags: [Areas, Dashboard]
      summary: Get area statistics
      description: Returns compliance score, task counts, alerts for an area
      parameters:
        - $ref: '#/components/parameters/AreaId'
      responses:
        '200':
          description: Area statistics
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AreaStats'

  # ============ USERS ============
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Clerk JWT token
  parameters:
    AreaId:
      name: areaId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
...
    Area:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        code:
          type: string
        parentId:
          type: string
          format: uuid
        responsibleId:
          type: string
          format: uuid
        responsibleName:
          type: string
        children:
          type: array
          items:
            $ref: '#/components/schemas/Area'
        sortOrder:
          type: integer
        createdAt:
          type: string
          format: date-time
    AreaWithStats:
      allOf:
        - $ref: '#/components/schemas/Area'
        - type: object
          properties:
            stats:
              $ref: '#/components/schemas/AreaStats'
    AreaStats:
      type: object
      properties:
        complianceScore:
          type: number
          minimum: 0
          maximum: 100
        tasksCompleted:
          type: integer
        tasksPending:
          type: integer
        tasksOverdue:
          type: integer
        openProblems:
          type: integer
        lastUpdated:
          type: string
          format: date-time
...
    CreateAreaRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          maxLength: 255
        code:
          type: string
          maxLength: 50
        parentId:
          type: string
          format: uuid
        responsibleId:
          type: string
          format: uuid
    UpdateAreaRequest:
      type: object
      properties:
        name:
          type: string
        code:
          type: string
        parentId:
          type: string
          format: uuid
        responsibleId:
          type: string
          format: uuid
        sortOrder:
          type: integer
...
    Error:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object